requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
celery==5.3.4
redis==5.0.1
deepgram-sdk==3.2.7
aiofiles==23.2.1
httpx==0.25.2
python-dotenv==1.0.0

